+++ src/Entity/Event.php
@@ -2,6 +2,7 @@
 
 namespace App\Entity;
 
+use App\Entity\Traits\SoftDeletableTrait;
 use App\Repository\EventRepository;
 use DateTime;
 use DateTimeInterface;
@@ -19,6 +20,8 @@ use Symfony\Component\Validator\Constraints as Assert;
  */
 class Event
 {
+    use SoftDeletableTrait;
+
     /**
      * @ORM\Id
      * @ORM\Column(type="uuid", unique=true)
@@ -103,13 +106,6 @@ class Event
      * @Assert\Type("\DateTimeInterface")
      */
     private $updatedAt;
-
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     *
-     * @Assert\Type("\DateTimeInterface")
-     */
-    private $deletedAt;
 
     /**
      * The relation to the EventCategory the Event is classified as.
@@ -272,23 +268,6 @@ class Event
         $this->updatedAt = $updatedAt;
 
         return $this;
-    }
-
-    public function getDeletedAt(): ?DateTimeInterface
-    {
-        return $this->deletedAt;
-    }
-
-    public function setDeletedAt(?DateTimeInterface $deletedAt): self
-    {
-        $this->deletedAt = $deletedAt;
-
-        return $this;
-    }
-
-    public function isSoftDeleted(): bool
-    {
-        return !(null === $this->deletedAt);
     }
 
     /**
