+++ src/Entity/UECourseExchange.php
@@ -2,6 +2,7 @@
 
 namespace App\Entity;
 
+use App\Entity\Traits\SoftDeletableTrait;
 use App\Repository\UECourseExchangeRepository;
 use DateTime;
 use DateTimeInterface;
@@ -20,6 +21,8 @@ use Symfony\Component\Validator\Constraints as Assert;
  */
 class UECourseExchange
 {
+    use SoftDeletableTrait;
+
     /**
      * @ORM\Id
      * @ORM\Column(type="uuid", unique=true)
@@ -89,13 +92,6 @@ class UECourseExchange
      * @Assert\Type("\DateTimeInterface")
      */
     private $updatedAt;
-
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     *
-     * @Assert\Type("\DateTimeInterface")
-     */
-    private $deletedAt;
 
     public function __construct()
     {
@@ -222,22 +218,5 @@ class UECourseExchange
         $this->updatedAt = $updatedAt;
 
         return $this;
-    }
-
-    public function getDeletedAt(): ?DateTimeInterface
-    {
-        return $this->deletedAt;
-    }
-
-    public function setDeletedAt(?DateTimeInterface $deletedAt): self
-    {
-        $this->deletedAt = $deletedAt;
-
-        return $this;
-    }
-
-    public function isSoftDeleted(): bool
-    {
-        return !(null === $this->deletedAt);
     }
 }
